Move transformers to detail directory
