+++ src/orange/CMakeLists.txt
@@ -30,8 +30,6 @@ list(APPEND SOURCES
   surf/SurfaceIO.cc
   surf/detail/SurfaceTranslator.cc
   surf/detail/SurfaceTransformer.cc
-  transform/TransformTransformer.cc
-  transform/TransformTranslator.cc
   transform/Transformation.cc
   transform/VariantTransform.cc
 )
